requirements.txt
setuptools
connexion[swagger-ui]
python_dateutil
mysql-connector-python
pandas